refactor: 폴더구조 변경 - root 디렉터리 생성
