+++ camel/week2/src/objects/MeshBasicMaterialObject.js
