+++ camel/week2/src/objects/PavingStones.js
