+++ camel/week2/textures/autumn_field_puresky_1k.hdr
