+++ camel/week2/textures/brown_photostudio_02_1k.hdr
